Migrate updated_dates().
